CATROID-1057 REFACTOR Look/PhysicsLook
Extracted PhysicsObjectStateHandler from PhysicsLook.
Reduce Code Duplication and Magic Number Usage, and general
Code Cleanup in PhysicsLook
