+++ catroid/src/main/java/org/catrobat/catroid/content/Look.java
@@ -786,7 +786,6 @@ public class Look extends Image {
 		}
 	}
 
-	@VisibleForTesting
 	public float getAlpha() {
 		return alpha;
 	}
